Add tests for song service

+++ app/main/services/songs.py
@@ -3,58 +3,77 @@ import datetime
 from ..models import db
 from ..models.songs import Songs
 
+from ..errors.errors import ValidationError, ServerError
+
 
 class SongService:
     def add(self, song_data):
+        if not 'name' in song_data or not 'duration' in song_data:
+            raise ValidationError
+
         new_song = Songs(
             name=song_data['name'],
             duration=song_data['duration'],
             uploaded_time=datetime.datetime.utcnow()
         )
 
-        self.__save(new_song)
-        response = {
-            'status': 'success',
-            'message': 'Song added successfully'
-        }
-
-        return response
+        try:
+            self.__save(new_song)
+            return {
+                'status': 'success',
+                'message': 'Song added successfully'
+            }
+        except Exception as e:
+            raise ServerError
 
     def get(self, id):
-        return Songs.query.filter_by(id=id).first()
+        song = Songs.query.filter_by(id=id).first()
+
+        if not song:
+            raise ValidationError
+
+        return song
 
     def getAll(self):
         return Songs.query.all()
 
     def edit(self, id, data):
         song = Songs.query.filter_by(id=id).first()
-        
+
+        if not song:
+            raise ValidationError
+
         if 'name' in data:
             song.name = data['name']
-            
+
         if 'duration' in data:
             song.duration = data['duration']
 
-        self.__save(song)
-        response = {
-            'status': 'success',
-            'message': 'Song edited successfully'
-        }
-
-        return response
+        try:
+            self.__save(song)
+            return {
+                'status': 'success',
+                'message': 'Song edited successfully'
+            }
+        except Exception as e:
+            raise ServerError
 
     def delete(self, id):
         song = Songs.query.filter_by(id=id).first()
 
-        db.session.delete(song)
-        db.session.commit()
+        if not song:
+            raise ValidationError
 
-        response = {
-            'status': 'success',
-            'message': 'Song deleted successfully'
-        }
+        try:
+            db.session.delete(song)
+            db.session.commit()
 
-        return response
+            return {
+                'status': 'success',
+                'message': 'Song deleted successfully'
+            }
+        except Exception as e:
+            raise ServerError
 
     def __save(self, data):
         db.session.add(data)
